Strategies setup service refactoring
