+++ libs/template/src/lib/strategies-setup/interfaces/strategies-setup-state.interface.ts
@@ -1,7 +1,7 @@
 import { ExtendedStrategies } from './extended-strategy';
 import { DefaultStrategyName } from './default-strategy-name.type';
 
-export interface StrategySetupState<S extends string = DefaultStrategyName> {
+export interface StrategiesSetupState<S extends string = DefaultStrategyName> {
   currentStrategy: keyof ExtendedStrategies<S>;
-  currentOutOfViewportStrategy: keyof ExtendedStrategies<S>;
+  currentInvisibleStrategy: keyof ExtendedStrategies<S>;
 }
